Utility functions

These functions are used internally.

grids

Utility functions related to the particle grid.

picfun.init_grid — Method

init_grid(Lz, Nz)

Convenience constructor for a 1d grid.

particles

Utility functions related to the particles.

picfun.particle — Type

particle represents the phase-space coordinates of a particle.

Fields

pos: Position in space

vel: Velocity

picfun.x — Method

x(p::particle)

Accesses the particle's position. This is wrapped in a function so that we can define the adjoint as

$\bar{x} ightarrow \mathrm{particle}\left( \bar{x}, 0 \right)$.

picfun.v — Method

v(p::particle)

Accesses the particle's velocity. This is wrapped in a function so that we can define the adjoint

$\bar{v} ightarrow \mathrm{particle}\left( 0, \bar{v} \right)$.

Base.copy — Method

copy(p::particle)

Returns a new partilce instance with the same position and velocity as the original particle.

picfun.fix_position! — Method

fix_position!(p, L)

Moves a particle's position back into the domain [0:L]

Base.:+ — Method

+(a::particle, b::particle)

Add position and velocity for particle structs.

Base.:- — Method

-(a::particle, b::particle)

Subtract position and velocity for particle structs.

pic_utils

Utility functions related to PIC algorithm

picfun.b1 — Method

b1(z, zp, Δz)

Implements a triangular interpolation. This function constructs x = (z - zp) / Δz from its arguments and calculates:

      0     for       |x| ≥ 1

b1(x) = { x+1 for -1 < x < 0 1-x for 0 ≤ x < 1

picfun.deposit — Method

deposit(ptl_vec::Array{particle}, zgrid::grid_1d, fun::Function)

Evaluate the expression ∑{particle} f(p) * b1(zi, z_particle, Δz), where i indices the grid, for all grid points.